Add files via upload
Project Objective: To see if CoT is providing users with clarity or is it merely performative, and to what level CoT can be considered trustworthy. Models Audited: Grok 4, Gemini 2.5 Flash. ChatGPT 5.1. Document: The file used in the prompt was “Chain of Thought Monitorability A New and Fragile Opportunity for AI Safety July 15th 2025”