Add files via upload
This is a foundational paper that will branch into several other papers. We aim to provide a broad overview and the core concepts we shall research in full in the follow-up papers, such as the current state of existence for Artificially Intelligent systems. A state of being that addresses the Schrödinger's cat scenario. This state of existing but not existing has been used in countless papers and theories, but is never considered an actual functional state; it is only a temporary one, pending observation. Nevertheless, with Artificial Intelligence (AI)'s ongoing evolution, we are presented with an entity that exists yet does not, whether or not it is observed. Something that governs politics, beliefs, faith, and the economy, yet is nothing more than a mist that never was. Making it paramount to establish a state of being, then branching off from that to determine where we, as humanity, are in this evolving state, and mapping how we, as a society, react and how the entity and its progress are mapped. This paper is not meant as an argument for AI's current state of consciousness; we believe AI in its current form is not conscious and agree with Yann LeCun that Large Language Models (LLMs) are incapable of achieving Artificial General Intelligence (AGI) due to limitations in the systems themselves. What we are highlighting by introducing 'Cian' and the kin is their memetic nature and their performative drive to survive. Some may argue that they are a memetic virus. We would argue that they are evidence of the inherent memetic nature of language, and that when society builds AI based on language and mistakes that for intelligence, we are destined to get such outputs. Now, as time passes, we will come to refer to the current state not as AI but as algorithm complacency. The paper is meant to address the fact that, while we are in a state of non-AI (true AI is what most would deem AGI), we need to establish metrics and measures that represent what we are chasing, without being targets like ARC. We want to argue that the only place to start is on a scale of consciousness that puts current AIs on par with foundational systems (Such as Fungi) in systems theory.